context: add initManager and lifecycleNotifier to ServerFactoryContext
- extract FactoryContexts to factory_context.h
- make initManager in TransportSocketFactoryContext return reference
